# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TelePhysiotherapy API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BlogPostType } from "./blogPostType";

export interface BlogPost {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  excerpt: string | null;
  author: string;
  type: BlogPostType;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  rating: number | null;
  publishedAt: string;
}