# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TelePhysiotherapy API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: number;
  name: string;
  email: string;
  phone: string;
  /** @nullable */
  age: number | null;
  /** @nullable */
  gender: string | null;
  /** @nullable */
  chiefComplaint: string | null;
  /** @nullable */
  medicalHistory: string | null;
  /** @nullable */
  painIntensity: number | null;
  consentGiven: boolean;
  createdAt: string;
}