# lib/api-zod/src/generated/types/progressEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TelePhysiotherapy API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressEntry {
  id: number;
  patientId: number;
  date: string;
  painIntensity: number;
  /** @nullable */
  notes: string | null;
  exercisesCompleted: boolean;
  createdAt: string;
}