# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TelePhysiotherapy API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceType } from "./serviceType";

export interface Service {
  id: number;
  name: string;
  description: string;
  type: ServiceType;
  /** @nullable */
  duration: string | null;
  /** @nullable */
  price: string | null;
  /** @nullable */
  imageUrl: string | null;
  features: string;
}