# lib/api-zod/src/generated/types/symptom.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TelePhysiotherapy API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Symptom {
  id: number;
  name: string;
  description: string;
  bodyPart: string;
  commonCauses: string;
  treatments: string;
  /** @nullable */
  imageUrl: string | null;
}