# lib/api-zod/src/generated/types/updateAppointmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TelePhysiotherapy API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateAppointmentBodyStatus } from "./updateAppointmentBodyStatus";

export interface UpdateAppointmentBody {
  /** @nullable */
  status?: UpdateAppointmentBodyStatus;
  /** @nullable */
  sessionLink?: string | null;
  /** @nullable */
  physiotherapist?: string | null;
  /** @nullable */
  notes?: string | null;
}